% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-figures.R
\name{base_tp_plot}
\alias{base_tp_plot}
\title{base_tp_plot}
\usage{
base_tp_plot(data, vpc_data, variable_names, variable_labels, use_tad = FALSE)
}
\arguments{
\item{data}{A data.frame of data}

\item{vpc_data}{A data.frame of VPC data}

\item{variable_names}{A list of variable names}

\item{variable_labels}{A list of variable labels}

\item{use_tad}{Logical, if \code{TRUE}, use TAD instead of time}
}
\description{
Base time profile plot
}
\keyword{internal}
