% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tables.R, R/utilities-vpc.R
\name{bin_values}
\alias{bin_values}
\title{bin_values}
\usage{
bin_values(values, bins = 7)

bin_values(values, bins = 7)
}
\arguments{
\item{values}{A vector of values to be binned}

\item{bins}{Number of bins to create}
}
\value{
A factor with the binned values

A factor with the binned values
}
\description{
Bin values into specified number of bins based on quantiles.

Bin values into specified number of bins based on quantiles.
}
\examples{

data_501 |> dplyr::mutate(Bins = bin_values(values = TIME, bins = 7))


data_501 |> dplyr::mutate(Bins = bin_values(values = TIME, bins = 7))

}
