% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem.R
\name{calc_tad}
\alias{calc_tad}
\title{calc_tad}
\usage{
calc_tad(id, time, amt = NULL, evid = NULL, rate = NULL, dur = NULL)
}
\arguments{
\item{id}{Subject identifiers values}

\item{time}{Time values}

\item{amt}{Amount values}

\item{evid}{Event identifiers values}

\item{rate}{Rate values}

\item{dur}{Duration values}
}
\description{
Calculate time after dose (TAD).
Records before the first dose are output < 0.
For infusions (\code{dur>0} or \code{rate>0}), the time after the end of infusion is calculated.
}
\examples{

data_501 |>
  dplyr::mutate(TAD2 = calc_tad(id = ID, time = TIME, amt = AMT)) |>
  dplyr::select(ID, TIME, AMT, TAD2)

data_501 |>
  dplyr::mutate(TAD2 = calc_tad(id = ID, time = TIME, evid = EVID)) |>
  dplyr::select(ID, TIME, EVID, TAD2)

data_501 |>
  dplyr::mutate(TAD2 = calc_tad(
    id = ID,
    time = TIME,
    evid = EVID,
    amt = AMT,
    rate = RATE
  )) |>
  dplyr::select(ID, TIME, EVID, AMT, RATE, TAD2)

}
