% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cat_inventory}
\alias{cat_inventory}
\title{cat_inventory}
\usage{
cat_inventory(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame summarizing the data
}
\description{
Get a data.frame summarizing categorical data
}
