% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{check_ranges}
\alias{check_ranges}
\title{check_ranges}
\usage{
check_ranges(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame summarizing the data
}
\description{
Get a data.frame summarizing data
}
