% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cor_report}
\alias{cor_report}
\title{cor_report}
\usage{
cor_report(data, x, y)
}
\arguments{
\item{data}{A data.frame}

\item{x}{name of \code{x} variable}

\item{y}{name of \code{y} variable}
}
\value{
A character of correlation and p-value
}
\description{
Report of Spearman correlation with p-value
}
