% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-inventory.R
\name{cov_inventory}
\alias{cov_inventory}
\title{cov_inventory}
\usage{
cov_inventory(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A list of data.frames summarizing the covariate data
}
\description{
Get a list of data.frames summarizing covariate data
}
\examples{

cov_inventories <- cov_inventory(data_501, meta_data_501)

cov_inventories$All

cov_inventories[["SEX : Female"]]

cov_inventories[["SEX : Male"]]

}
