% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem-runs.R
\name{cov_to_cor}
\alias{cov_to_cor}
\title{cov_to_cor}
\usage{
cov_to_cor(data, cov_name = "Estimates")
}
\arguments{
\item{data}{A data.frame}

\item{cov_name}{Name of the column containing covariance values}
}
\description{
Translate covariance to correlation values
}
