% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{data_501}
\alias{data_501}
\title{data_501}
\format{
An object of class \code{data.frame} with 300 rows and 12 columns.
}
\usage{
data_501
}
\description{
Dataset from Nonmem 501 example
}
\keyword{datasets}
