% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{data_inventory}
\alias{data_inventory}
\title{data_inventory}
\usage{
data_inventory(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A list of data.frame summarizing the data
}
\description{
Get a data.frame summarizing data
}
