% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-meta-data.R
\name{dictionary_check}
\alias{dictionary_check}
\title{dictionary_check}
\usage{
dictionary_check(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame with the meta data variables filled in
}
\description{
Check if consistency between \code{data} and \code{meta_data}
}
