% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{dplyrVariables}
\alias{dplyrVariables}
\title{dplyrVariables}
\format{
An object of class \code{character} of length 54.
}
\usage{
dplyrVariables
}
\description{
Remove check warning for the variables listed below
Variables called withing dplyr and tidyr functions
}
\keyword{internal}
