% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-gof.R
\name{dv_ipred}
\alias{dv_ipred}
\title{dv_ipred}
\usage{
dv_ipred(data, meta_data = NULL)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot DV vs IPRED
}
\examples{

dv_ipred(data_501, meta_data_501)

}
