% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-gof.R
\name{dv_pred}
\alias{dv_pred}
\title{dv_pred}
\usage{
dv_pred(data, meta_data = NULL)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot DV vs PRED
}
\examples{

dv_pred(data_501, meta_data_501)

}
