% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{dv_preds}
\alias{dv_preds}
\title{dv_preds}
\usage{
dv_preds(data, meta_data = NULL)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot DV vs PREDs
}
