% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{eta_cor}
\alias{eta_cor}
\title{eta_cor}
\usage{
eta_cor(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame summarizing the data
}
\description{
Get a data.frame of
Spearman correlation between covariates and etas
Wilcoxon test between categorical covariates and etas
}
