% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-covariate.R
\name{eta_plot}
\alias{eta_plot}
\title{eta_plot}
\usage{
eta_plot(data, meta_data = NULL)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Pairs plot of etas
}
\examples{

eta_plot(data_501, meta_data_501)

}
