% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-meta-data.R
\name{fill_meta_vars}
\alias{fill_meta_vars}
\title{fill_meta_vars}
\usage{
fill_meta_vars(meta_data = NULL)
}
\arguments{
\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame with the meta data variables filled in
}
\description{
Fill required meta data types and variables
}
