% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-meta-data.R
\name{fill_nonmem_vars}
\alias{fill_nonmem_vars}
\title{fill_nonmem_vars}
\usage{
fill_nonmem_vars(data, lloq = 10, group = "All")
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{lloq}{Lower limit of quantification (default is 10)}

\item{group}{Name of the grouping variable}
}
\value{
A data.frame with the nonmem variables filled in
}
\description{
Fill in the nonmem variables that are not present in the data
}
