% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem-runs.R
\name{find_row}
\alias{find_row}
\title{find_row}
\usage{
find_row(patterns, text)
}
\arguments{
\item{patterns}{character vector of patterns to match}

\item{text}{character vector of text to search}
}
\description{
Find row that matches patterns
}
\keyword{internal}
