% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem.R
\name{format_result}
\alias{format_result}
\title{format_result}
\usage{
format_result(nonmem_result)
}
\arguments{
\item{nonmem_result}{A list of Nonmem results}
}
\description{
Format results as a data.frame
}
