% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{gg_lim}
\alias{gg_lim}
\title{gg_lim}
\usage{
gg_lim(plot_object, meta_data, x = NULL, y = NULL)
}
\arguments{
\item{plot_object}{A ggplot object}

\item{meta_data}{A meta_data data.frame}

\item{x}{A variable name for x-axis}

\item{y}{A variable name for y-axis}
}
\description{
Apply \code{Min} and \code{Max} from \code{meta_data} as limits of the ggplot
}
