% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-figures.R
\name{gg_log}
\alias{gg_log}
\title{gg_log}
\usage{
gg_log(plot_object, x = TRUE, y = TRUE)
}
\arguments{
\item{plot_object}{A ggplot object}

\item{x}{Logical, if \code{TRUE}, x-axis will be log scale}

\item{y}{Logical, if \code{TRUE}, y-axis will be log scale}
}
\description{
Use pretty log scale plot
}
