% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{highlight_significant}
\alias{highlight_significant}
\title{highlight_significant}
\usage{
highlight_significant(data, pval = 0.05, color = "green", format = "html")
}
\arguments{
\item{data}{A data.frame}

\item{pval}{Significance threshold for p-value}

\item{color}{Color when significant}

\item{format}{\code{"typst"}, \code{"html"}}
}
\description{
Add color for significant cell
}
\examples{

cor_data <- cov_cor(data_501, meta_data_501)
cor_data

cor_data |> highlight_significant(pval = 0.1, format = "html")

cor_data |> highlight_significant(pval = 0.1, format = "typst")

cor_data |> highlight_significant(pval = 0.1, format = "docx")

}
\seealso{
Other report: 
\code{\link{report_dataset_analysis}()},
\code{\link{report_gof_analysis}()},
\code{\link{report_limited_sampling_analysis}()},
\code{\link{report_vpc_analysis}()},
\code{\link{report_workflow_summary}()},
\code{\link{typst_scientific}()},
\code{\link{typst_table}()}
}
\concept{report}
