% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-time-profile.R
\name{ind_time_profiles}
\alias{ind_time_profiles}
\title{ind_time_profiles}
\usage{
ind_time_profiles(data, meta_data = NULL, n_rows = 2, n_cols = 3)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}

\item{n_rows}{Number of rows in the plot grid, default is 2}

\item{n_cols}{Number of columns in the plot grid, default is 3}
}
\description{
Plot DV, PRED and IPRED vs TIME
}
\examples{

# Simulate 1-compartment model
pk_data <- data_501 |>
  dplyr::mutate(CL = 2, V = 40)
ind_time_profiles(pk_data, meta_data_501)

}
