% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{lm_report}
\alias{lm_report}
\title{lm_report}
\usage{
lm_report(data, x, y)
}
\arguments{
\item{data}{A data.frame}

\item{x}{name of \code{x} variable}

\item{y}{name of \code{y} variable}
}
\value{
A character of lm coefficients and p-value
}
\description{
Report of lm coefficients with p-value
}
