% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem-runs.R
\name{matrix_indices}
\alias{matrix_indices}
\title{matrix_indices}
\usage{
matrix_indices(n_dim)
}
\arguments{
\item{n_dim}{Matrix dimension}
}
\description{
Get x and y indices of a \code{n_dim}x\code{n_dim} matrix
}
\keyword{internal}
