% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{meta_data_501}
\alias{meta_data_501}
\title{data_501}
\format{
An object of class \code{data.frame} with 12 rows and 6 columns.
}
\usage{
meta_data_501
}
\description{
Dictionary example for Nonmem 501 example dataset
}
\keyword{datasets}
