% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem-runs.R
\name{parse_in_text}
\alias{parse_in_text}
\title{parse_in_text}
\usage{
parse_in_text(patterns, text, type = "numeric")
}
\arguments{
\item{patterns}{character vector of patterns to match}

\item{text}{character vector of text to search}

\item{type}{type of variable to parse}
}
\description{
Parse a variable at the end of a line matching target patterns
}
\keyword{internal}
