% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem.R
\name{parse_omega}
\alias{parse_omega}
\title{parse_omega}
\usage{
parse_omega(n_omega, text)
}
\arguments{
\item{text}{character vector of text to search}

\item{n_sigma}{Number of SIGMA parameters}
}
\description{
Parse OMEGA (random effects - etas) characteristics in text
}
\keyword{internal}
