% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem.R
\name{parse_omega_size}
\alias{parse_omega_size}
\title{parse_omega}
\usage{
parse_omega_size(text)
}
\arguments{
\item{text}{character vector of text to search}
}
\description{
Parse OMEGA size when BOV is used and the matrix is using blocks
}
\keyword{internal}
