% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{parse_pvalue}
\alias{parse_pvalue}
\title{parse_pvalue}
\usage{
parse_pvalue(text)
}
\arguments{
\item{text}{character vector}
}
\description{
Parse pvalue as \code{"(p: 0.xyz)"} or \code{"(p< 0.001)"} from text
}
\keyword{internal}
