% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem-runs.R
\name{parse_sigma}
\alias{parse_sigma}
\title{parse_sigma}
\usage{
parse_sigma(n_sigma, text)
}
\arguments{
\item{n_sigma}{Number of SIGMA parameters}

\item{text}{character vector of text to search}
}
\description{
Parse SIGMA (random effects - eps) characteristics in text
}
\keyword{internal}
