% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem.R
\name{parse_theta}
\alias{parse_theta}
\title{parse_theta}
\usage{
parse_theta(n_theta, text)
}
\arguments{
\item{n_theta}{Number of THETA parameters}

\item{text}{character vector of text to search}
}
\description{
Parse THETA (fixed effects) characteristics in text
}
\keyword{internal}
