% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-figures.R
\name{plotly_log}
\alias{plotly_log}
\title{plotly_log}
\usage{
plotly_log(plotly_object, x = TRUE, y = TRUE)
}
\arguments{
\item{plotly_object}{A plotly object}

\item{x}{Logical, if \code{TRUE}, x-axis will be log scale}

\item{y}{Logical, if \code{TRUE}, y-axis will be log scale}
}
\description{
Use log scale when plotting with \code{plotly}
}
