% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-meta-data.R
\name{pull_label}
\alias{pull_label}
\title{pull_label}
\usage{
pull_label(name, meta_data = NULL)
}
\arguments{
\item{name}{Name of variable to retrieve}

\item{meta_data}{A data.frame of meta data}
}
\description{
Pull Label potentially with Unit from \code{Name} field
}
\examples{
# Dictionary
meta_data_501

pull_label("TIME", meta_data_501)

pull_label(c("WT", "AGE", "SEX"), meta_data_501)

}
