% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-meta-data.R
\name{pull_limits}
\alias{pull_limits}
\title{pull_limits}
\usage{
pull_limits(name, meta_data = NULL)
}
\arguments{
\item{name}{Name of variable to retrieve}

\item{meta_data}{A data.frame of meta data}
}
\description{
Pull limits \code{Name} field
}
\examples{
# Dictionary
meta_data_501

pull_limits("TIME", meta_data_501)

}
