% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{report_dataset_analysis}
\alias{report_dataset_analysis}
\title{report_dataset_analysis}
\usage{
report_dataset_analysis(
  data_path,
  meta_data_path,
  report_path = "dataset-analysis.docx",
  bins = 7
)
}
\arguments{
\item{data_path}{Path to the data file}

\item{meta_data_path}{Path to the metadata file}

\item{report_path}{Path to the output report file}

\item{bins}{Number of bins for time profile plots}
}
\description{
Render a docx document that analyses a dataset
}
\seealso{
Other report: 
\code{\link{report_gof_analysis}()},
\code{\link{report_limited_sampling_analysis}()},
\code{\link{report_vpc_analysis}()},
\code{\link{report_workflow_summary}()}
}
\concept{report}
