% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{report_gof_analysis}
\alias{report_gof_analysis}
\title{report_gof_analysis}
\usage{
report_gof_analysis(
  data_path,
  meta_data_path,
  report_path = "gof-analysis.docx"
)
}
\arguments{
\item{data_path}{Path to the data file}

\item{meta_data_path}{Path to the metadata file}

\item{report_path}{Path to the output report file}
}
\description{
Render a docx document that analyses goodness of fit results
}
\seealso{
Other report: 
\code{\link{report_dataset_analysis}()},
\code{\link{report_limited_sampling_analysis}()},
\code{\link{report_vpc_analysis}()},
\code{\link{report_workflow_summary}()}
}
\concept{report}
