% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{report_limited_sampling_analysis}
\alias{report_limited_sampling_analysis}
\title{report_limited_sampling_analysis}
\usage{
report_limited_sampling_analysis(
  model_path,
  data_path,
  sampling_path = "Sampling-Strategies.xlsx",
  run_folder = "LSA_Runs",
  pop_size = 1000,
  pop_cv = 0.025,
  covariates = c("BW", "AGE", "FFM"),
  categoricals = NULL,
  lloq = 0.01,
  baseline = 0.005,
  y_lim = 50,
  report_path = "limited-sampling-analysis.docx"
)
}
\arguments{
\item{model_path}{Path to the \code{.cpp} model file}

\item{data_path}{Path to the data file}

\item{sampling_path}{Path to the sampling strategies \code{.xslx} file.
Template file available at \code{system.file("Sampling-Strategies.xlsx", package = "nonmem.utils")}}

\item{run_folder}{Path to the folder where the runs will be stored}

\item{pop_size}{Population size for the analysis}

\item{pop_cv}{Coefficient of variation for the population}

\item{covariates}{Covariate names to be used in the analysis}

\item{categoricals}{Categorical covariate names to be used in the analysis}

\item{lloq}{Lower limit of quantification for the analysis}

\item{baseline}{Baseline value for the analysis}

\item{y_lim}{Y-axis limits for the plots}

\item{report_path}{Path to the output report file}
}
\description{
Render a docx document that report limited sampling analyses results
}
\seealso{
Other report: 
\code{\link{report_dataset_analysis}()},
\code{\link{report_gof_analysis}()},
\code{\link{report_vpc_analysis}()},
\code{\link{report_workflow_summary}()}
}
\concept{report}
