% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{report_workflow_summary}
\alias{report_workflow_summary}
\title{report_workflow_summary}
\usage{
report_workflow_summary(dir_path, report_path = "workflow-summary.docx")
}
\arguments{
\item{dir_path}{Path to the nonmem results directory}

\item{report_path}{Path to the output report file}
}
\description{
Render a docx document that analyses the nonmem results
}
\seealso{
Other report: 
\code{\link{highlight_significant}()},
\code{\link{report_dataset_analysis}()},
\code{\link{report_gof_analysis}()},
\code{\link{report_limited_sampling_analysis}()},
\code{\link{report_vpc_analysis}()},
\code{\link{typst_scientific}()},
\code{\link{typst_table}()}
}
\concept{report}
