% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{residual_hist}
\alias{residual_hist}
\title{residual_plot}
\usage{
residual_hist(y_type = "cwres", data, meta_data = NULL)
}
\arguments{
\item{y_type}{Type of y variable}

\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot histogram of residual (\code{y})
}
