% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-gof.R
\name{residual_plot}
\alias{residual_plot}
\title{residual_plot}
\usage{
residual_plot(x_type = "time", y_type = "cwres", data, meta_data = NULL)
}
\arguments{
\item{x_type}{Type of x variable}

\item{y_type}{Type of y variable}

\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot residual (\code{y}) vs \code{x}
}
\examples{

residual_plot(
x_type = "time",
y_type = "cwres",
data = data_501,
meta_data = meta_data_501
)

}
