% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-gof.R
\name{residual_qq}
\alias{residual_qq}
\title{residual_qq}
\usage{
residual_qq(y_type = "cwres", data, meta_data = NULL)
}
\arguments{
\item{y_type}{Type of y variable}

\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot qq-plot of residuals (\code{y})
}
\examples{

residual_qq(
y_type = "cwres",
data = data_501,
meta_data = meta_data_501
)

}
