% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tables.R, R/utilities-vpc.R
\name{run_vpc}
\alias{run_vpc}
\title{run_vpc}
\usage{
run_vpc(data, x = "TIME", group = "All", bins = 7, ci = 0.8, lloq = 10)

run_vpc(data, x = "TIME", group = "All", bins = 7, ci = 0.8, lloq = 10)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{x}{Name of \code{x} variable, usually \code{"TIME"}}

\item{group}{Name of \code{group} variable}

\item{bins}{Number of bins for the VPC}

\item{ci}{Confidence interval for the VPC, usually \code{0.8} (10th-90th percentiles)}

\item{lloq}{Lower limit of quantification}
}
\value{
A data.frame

A data.frame
}
\description{
Run a VPC analysis on the provided data.

Run a VPC analysis on the provided data.
}
\examples{

run_vpc(data_501, x = "TIME")


run_vpc(data_501, x = "TIME")

}
