% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem.R
\name{simulate_pk}
\alias{simulate_pk}
\title{simulate_pk}
\usage{
simulate_pk(data)
}
\arguments{
\item{data}{A data.frame of PK data.}
}
\value{
A data.frame with simulated PK data.
}
\description{
Simulate PK from PK data.
Currently, only 1-, 2- and 3-compartment infusion models are supported.
}
\examples{

head(simulate_pk(data_501 |> dplyr::filter(ID == 1)))
}
