% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{tad_profile}
\alias{tad_profile}
\title{tad_profile}
\usage{
tad_profile(data, meta_data = NULL, bins = 7)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot Time After Dose (TAD) vs DV
}
