% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{time_profile}
\alias{time_profile}
\title{time_profile}
\usage{
time_profile(data, meta_data = NULL, bins = 7)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot TIME vs DV
}
