% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-figures.R
\name{tooltip_text}
\alias{tooltip_text}
\title{tooltip_text}
\usage{
tooltip_text(data, var_names)
}
\arguments{
\item{data}{A data.frame of data}

\item{var_names}{A vector of variable names to display}
}
\description{
When interactive plots are generated using \code{plotly},
\code{tootltip} will display text aesthetic allowing easy tracking of data.
}
