% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{typst_scientific}
\alias{typst_scientific}
\title{typst_scientific}
\usage{
typst_scientific(values, digit = 2)
}
\arguments{
\item{values}{Values to render into scientific format}

\item{digit}{Number of digits}
}
\description{
Add characters to render number into scientific writing
}
\examples{

typst_scientific(stats::rlnorm(5))

}
\seealso{
Other report: 
\code{\link{highlight_significant}()},
\code{\link{report_dataset_analysis}()},
\code{\link{report_gof_analysis}()},
\code{\link{report_limited_sampling_analysis}()},
\code{\link{report_vpc_analysis}()},
\code{\link{report_workflow_summary}()},
\code{\link{typst_table}()}
}
\concept{report}
