% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-rendering.R
\name{typst_table}
\alias{typst_table}
\title{typst_table}
\usage{
typst_table(data, linebreak = "<br>")
}
\arguments{
\item{data}{A data.frame to render}

\item{linebreak}{Line break pattern to be subbed by typst line break}
}
\description{
Print a table in raw typst for pdf document rendering
}
\examples{

data_with_linebreak <- data.frame(Example = c("$10^2$<br>$-->$", "#sym.checkmark"))
typst_table(data_with_linebreak)

}
\seealso{
Other report: 
\code{\link{highlight_significant}()},
\code{\link{report_dataset_analysis}()},
\code{\link{report_gof_analysis}()},
\code{\link{report_limited_sampling_analysis}()},
\code{\link{report_vpc_analysis}()},
\code{\link{report_workflow_summary}()},
\code{\link{typst_scientific}()}
}
\concept{report}
