% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{vpc_summary}
\alias{vpc_summary}
\title{vpc_summary}
\usage{
vpc_summary(data, x, y, group = NULL, bins = 5, stairstep = FALSE, ci = 0.9)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{x}{Name of \code{x} variable}

\item{y}{Name of \code{y} variable}

\item{group}{Name of \code{group} variable}

\item{bins}{Number of bins if one value is provided.
Or edges of bins if an array is provided.
Default binning uses number of rows/10}

\item{stairstep}{Logical defining if VPC data.frame correspond to stairstep plot}

\item{ci}{Confidence interval (value between 0 and 1)}
}
\value{
A data.frame summarizing the data distribution
with variables: \code{bins}, \code{x}, \code{n}, \code{y}, \code{ymin}, \code{ymax}, and \code{blq}
}
\description{
Get a data.frame summarizing data as a vpc
}
