% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bin_values}
\alias{bin_values}
\title{bin_values}
\usage{
bin_values(values, bins = 7)
}
\arguments{
\item{values}{A vector of values to be binned}

\item{bins}{Number of bins to create}
}
\value{
A factor with the binned values
}
\description{
Bin values into specified number of bins based on quantiles.
}
