% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{boot_hist}
\alias{boot_hist}
\title{boot_hist}
\usage{
boot_hist(y_type = "theta", data, meta_data = NULL, bins = 11, ci = 0.95)
}
\arguments{
\item{y_type}{Type of y variable}

\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}

\item{bins}{Number of bins}

\item{ci}{Confidence interval}
}
\description{
Histogram of bootstrap estimated (\code{y})
}
