% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cov_cor}
\alias{cov_cor}
\title{cov_cor}
\usage{
cov_cor(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame summarizing the data
}
\description{
Get a data.frame of Spearman correlation between covariates
}
\examples{

# Summarize data by ID
sum_data <- data_501 |>
  dplyr::group_by(ID) |>
  dplyr::summarise_all(dplyr::first)

cov_cor(sum_data, meta_data_501)

}
