% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cov_inventory}
\alias{cov_inventory}
\title{cov_inventory}
\usage{
cov_inventory(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame summarizing the data
}
\description{
Get a data.frame summarizing covariate data
}
