% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{cov_plot}
\alias{cov_plot}
\title{cov_plot}
\usage{
cov_plot(data, meta_data = NULL)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Pairs plot of covariates
}
