% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem.R
\name{cov_to_cor}
\alias{cov_to_cor}
\title{cov_to_cor}
\usage{
cov_to_cor(data, cov_name = "Estimates")
}
\arguments{
\item{data}{A data.frame}
}
\description{
Translate covariance to correlation values
}
