% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{data_inventory}
\alias{data_inventory}
\title{data_inventory}
\usage{
data_inventory(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A list of data.frame summarizing the data
}
\description{
Get a list of data.frames summarizing data
}
\examples{

# Currently requires BLQ data
inventory_meta <- dplyr::bind_rows(
  meta_data_501,
  data.frame(Name = "BLQ", Type = "blq", Label = "Below LLOQ")
)
inventory_data <- data_501 |> dplyr::mutate(BLQ = 0)

all_inventories <- data_inventory(inventory_data, inventory_meta)

all_inventories$All

all_inventories$Sex

}
