% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-inventory.R
\name{data_inventory}
\alias{data_inventory}
\title{data_inventory}
\usage{
data_inventory(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A list of data.frame summarizing the data
}
\description{
Get a list of data.frames summarizing data
}
\examples{

all_inventories <- data_inventory(data_501, meta_data_501)

# Since meta_data_501 identifies SEX as cat covariate
# inventory has a field split by Sex categories
names(all_inventories)

all_inventories$All

all_inventories$Sex


}
