% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-correlation.R
\name{eta_cor}
\alias{eta_cor}
\title{eta_cor}
\usage{
eta_cor(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\value{
A data.frame summarizing the data
}
\description{
Get a data.frame of
\itemize{
\item Spearman correlation between covariates and etas
\item Wilcoxon test between categorical covariates and etas
}
}
\examples{

# Summarize data by ID
sum_data <- data_501 |>
  dplyr::group_by(ID) |>
  dplyr::summarise_all(dplyr::first)

eta_cor(sum_data, meta_data_501)

}
