% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-covariate.R
\name{eta_cov_plot}
\alias{eta_cov_plot}
\title{eta_cov_plot}
\usage{
eta_cov_plot(data, meta_data = NULL)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Plot of etas vs covariates
}
\examples{

eta_cov_plot(data_501, meta_data_501)

}
