% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-meta-data.R
\name{map_cat_data}
\alias{map_cat_data}
\title{map_cat_data}
\usage{
map_cat_data(data, meta_data)
}
\arguments{
\item{data}{A data.frame of the pk data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Map categorical factor levels to a data.frame
}
