% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-nonmem-runs.R
\name{parse_omega}
\alias{parse_omega}
\title{parse_omega}
\usage{
parse_omega(n_omega, text)
}
\arguments{
\item{n_omega}{Number of OMEGA parameters}

\item{text}{character vector of text to search}
}
\description{
Parse OMEGA (random effects - etas) characteristics in text
}
\keyword{internal}
