% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-meta-data.R
\name{pull_name}
\alias{pull_name}
\title{pull_name}
\usage{
pull_name(type, meta_data = NULL)
}
\arguments{
\item{type}{Type of variable to retrieve}

\item{meta_data}{A data.frame of meta data}
}
\description{
Pull Name from data \code{Type} field
}
\examples{
# Dictionary
meta_data_501

pull_name("time", meta_data_501)

pull_name("cov", meta_data_501)

}
