% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-gof.R
\name{residual_hist}
\alias{residual_hist}
\title{residual_plot}
\usage{
residual_hist(y_type = "cwres", data, meta_data = NULL, bins = 21)
}
\arguments{
\item{y_type}{Type of y variable}

\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}

\item{bins}{Number of bins}
}
\description{
Plot histogram of residual (\code{y})
}
\examples{

residual_hist(
y_type = "cwres",
data = data_501,
meta_data = meta_data_501,
bins = 11
)

}
