% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{vpc_plots}
\alias{vpc_plots}
\title{vpc_plots}
\usage{
vpc_plots(data, meta_data, x = "TAD", bins = 7, ci = 0.8, lloq = 10)
}
\arguments{
\item{data}{A data.frame of data}

\item{meta_data}{A data.frame of meta data}
}
\description{
Generate VPC plots for a given dataset.
}
